+++ R/makeUR.R
@@ -30,8 +30,11 @@
 #'
 #' The filtering criteria currently implemented are
 #' \itemize{
-#' \item{Minor allele frequency (MAF): }{SNPs are discarded if their MAF is less than the threshold (default is 0.05)}
-#' \item{Proportion of missing data (MISS): }{SNPs are discarded if the proportion of individuals with no reads (e.g. missing genotype) is greater than the threshold value (default is 0.5)}
+#' \item{Minor allele frequency (MAF): }{SNPs are discarded if their MAF is less than the threshold (default is 0.01)}
+#' \item{Proportion of missing data (MISS): }{SNPs are discarded if the proportion of individuals with no reads
+#' (e.g. missing genotype) is greater than the threshold value (default is 0.5)}
+#' \item{Maximum average SNP read depth (MAXDEPTH): }{SNPs are discarded if the average read depth for the SNP
+#' is larger than the threshold (default is 500)}
 #' }
 #'
 #' Estimation of the allele frequencies when \code{mafEst=TRUE} is parallelized using the \code{\link[foreach]{foreach}} function, where the
@@ -39,10 +42,12 @@
 #' set the number of cores to be more than what is available on your computer (or bad things will happen!).
 #'
 #' @param RAobj Object of class RA created via the \code{\link{readRA}} function.
-#' @param filter Named list of thresholds for various criteria used to fiter SNPs.
-#' See below for details.
+#' @param indsubset Integer vector specifying which samples of the RA dataset to retain in the UR
+#' population.
 #' @param ploid An integer number specifying the ploidy level of the population. Currently, only
 #' even ploidy levels are valid.
+#' @param filter Named list of thresholds for various criteria used to fiter SNPs.
+#' See below for details.
 #' @param mafEst Logical value indicating whether the allele frequences and sequencing
 #' error parameters are to estimated for each SNP (see details).
 #' @param nClust Integer vector specifying the number of clusters to use in the foreach loop. Only used in the estimation of
@@ -58,27 +63,39 @@
 #' simdata <- readRA(RAfile)
 #'
 #' ## make unrelated population
-#' urpop <- makeUR(simdata, mafEst=FALSE, nClust=1)
+#' urpop <- makeUR(simdata, mafEst=FALSE)
 
 #### Make an unrelated population
-makeUR <- function(RAobj, filter=list(MAF=0.05, MISS=0.5), ploid=2, mafEst=TRUE, nClust=3){
+makeUR <- function(RAobj, indsubset=NULL, ploid=2, filter=list(MAF=0.01, MISS=0.5, MAXDEPTH=500), mafEst=TRUE, nClust=2){
 
   ## Do some checks
   if(!all(class(RAobj) %in% c("RA","R6")))
-    stop("First argument supplied is not of class 'R6' and 'RA'")
-  if(is.null(filter$MAF)) filter$MAF <- 0.05
+    stop("The `RAobj` argument supplied is not of class 'R6' and 'RA'")
+  if(is.null(filter$MAF)) filter$MAF <- 0.01
   else if( length(filter$MAF) != 1 || !is.numeric(filter$MAF) || filter$MAF<0 || filter$MAF>1)
     stop("Minor allele frequency filter is invalid")
   if(is.null(filter$MISS)) filter$MISS <- 0.2
   else if( length(filter$MISS) != 1 || !is.numeric(filter$MISS) || filter$MISS<0 || filter$MISS>1 )
     stop("Proportion of missing data filter is invalid")
+  if(is.null(filter$MAXDEPTH)) filter$MAXDEPTH <- 500
+  else if(checkVector(filter$MAXDEPTH, type="pos_numeric", minv=0, equal=FALSE) || length(filter$MAXDEPTH) != 1)
+    stop("Maximum mean SNP depth filter is invalid.")
+  #if(is.null(filter$PVALUE)) filter$PVALUE <- 1e-6
+  #else if( length(filter$PVALUE) != 1 || !is.numeric(filter$PVALUE) || filter$PVALUE<0 || filter$PVALUE>1 )
+   #stop("P-value for Hardy-Weinberg equilibrium filter is invalid.")
   if(!is.vector(ploid) || !is.numeric(ploid) || length(ploid) != 1 || round(ploid/2) != ploid/2)
     stop("Argument for ploid level is invalid.")
   if(!is.numeric(nClust) || length(nClust) != 1 || nClust < 0 || round(nClust) != nClust)
     stop("Argument for the number of cores for the parallelization is invalid")
-  #if(is.null(filter$HWdis)) filter$HWdis <- c(-0.05, 1)
-  #else if(!is.vector(filter$HWdis) || length(filter$HWdis) != 2 || !is.numeric(filter$HWdis) || filter$HWdis<0 || filter$HWdis >1)
-  #  stop("Hardy Weinberg Equilibrium (HWE) filter is invalid")
+  if(is.null(indsubset)) indsubset <- 1:RAobj$.__enclos_env__$private$nInd
+  else if(!is.vector(indsubset) || !is.character(indsubset) || any(is.na(indsubset)) ||
+          any(!(indsubset %in% RAobj$.__enclos_env__$private$indID)))
+    stop("Error in `indsubset` argument. At least one sample name was not found.")
+  else indsubset <- match(indsubset, RAobj$.__enclos_env__$private$indID)
+  if(checkVector(mafEst, type="one_logical"))
+    stop("The `mafEst` argument needs to be a single logical value.")
+  #if(checkVector(err, type="one_logical"))
+  #  stop("The `err` argument needs to be a single logical value.")
 
   ## initalize the UR object
   URobj <- UR$new(RAobj, ploid)
@@ -92,51 +109,42 @@ makeUR <- function(RAobj, filter=list(MAF=0.05, MISS=0.5), ploid=2, mafEst=TRUE,
   #cat("Hardy-Weinberg equilibrium: < ", filter$HWdis[1]," and > ",filter$HWdis[2],"\n\n",sep="")
 
   ## Extract the private variables we want
-  indID <- URobj$.__enclos_env__$private$indID
+  indsubset <- sort(unique(indsubset))
+  indID <- URobj$.__enclos_env__$private$indID[indsubset]
   nSnps <- URobj$.__enclos_env__$private$nSnps
-  genon <- URobj$.__enclos_env__$private$genon
-  ## Calculate the MAF
+  genon <- URobj$.__enclos_env__$private$genon[indsubset,]
+
+  ## Calculate the proportion of missing data and mean depths
+  miss <- apply(genon,2, function(x) sum(is.na(x))/length(x))
+  mdepth <- colMeans(URobj$.__enclos_env__$private$ref[indsubset,] + URobj$.__enclos_env__$private$alt[indsubset,])
+  ## do prelimiary filtering
+  snpsubset <- which(miss < filter$MISS & mdepth < filter$MAXDEPTH)
+
+  ## estimate allele frequencies and sequencing error parameters
   if(mafEst){
-    temp <- URobj$.__enclos_env__$private$p_est(nClust=nClust)
+    temp <- URobj$.__enclos_env__$private$p_est(snpsubset=snpsubset, indsubset=indsubset, nClust=nClust)
     pfreq <- unname(temp[1,])
     ep <- unname(temp[2,])
-    #ll_HWE <- unname(temp[3,])
-    #temp <- URobj$.__enclos_env__$private$g_est(nClust=nClust, err=err, para=list(ep=0))
-    #ll_gest <- temp[ploid+2,]
-    #hwe_pvalue <- 1-pchisq(-2*(ll_HWE - ll_gest), df=ploid-1)
   }
   else{
-    ratio <- URobj$.__enclos_env__$private$ref/(URobj$.__enclos_env__$private$ref+URobj$.__enclos_env__$private$alt)
+    ratio <- URobj$.__enclos_env__$private$ref[indsubset,snpsubset]/(URobj$.__enclos_env__$private$ref[indsubset,snpsubset]+URobj$.__enclos_env__$private$alt[indsubset,snpsubset])
     pfreq <- colMeans(ratio, na.rm=T)
-    ep <- rep(0, nSnps)
+    ep <- rep(0, length(snpsubset))
   }
-
-  ## Calculate the proportion of missing data
-  miss <- apply(genon,2, function(x) sum(is.na(x))/length(x))
-
-  # ## Compute the HWE distance
-  # naa <- colSums(genon == 2, na.rm = TRUE)
-  # nab <- colSums(genon == 1, na.rm = TRUE)
-  # nbb <- colSums(genon == 0, na.rm = TRUE)
-  # n1 <- 2 * naa + nab
-  # n2 <- nab + 2 * nbb
-  # n <- n1 + n2  #n alleles
-  # p1 <- n1/n
-  # p2 <- 1 - p1
-  # HWdis <- naa/(naa + nab + nbb) - p1 * p1
-
-  ## Indx the filtered SNPs
+  ## Compute MAF
   maf <- pmin(pfreq,1-pfreq)
-  indx <- (maf > filter$MAF) & (miss < filter$MISS) #& (HWdis > filter$HWdis[1]) & (HWdis < filter$HWdis[2])
+  ## subset SNPs
+  maf_indx <- which(maf > filter$MAF)
+  indx <- snpsubset[maf_indx]
 
   ## Update the data in the R6 object
   genon <- genon[,indx]
-  ref <- URobj$.__enclos_env__$private$ref[,indx]
-  alt <- URobj$.__enclos_env__$private$alt[,indx]
+  ref <- URobj$.__enclos_env__$private$ref[indsubset,indx]
+  alt <- URobj$.__enclos_env__$private$alt[indsubset,indx]
   SNP_Names <- URobj$.__enclos_env__$private$SNP_Names[indx]
-  nSnps <- sum(indx)
-  pfreq <- pfreq[indx]
-  ep <- ep[indx]
+  nSnps <- length(indx)
+  pfreq <- pfreq[maf_indx]
+  ep <- ep[maf_indx]
   if(URobj$.__enclos_env__$private$gform == "reference"){
     chrom = URobj$.__enclos_env__$private$chrom[indx]
     pos = URobj$.__enclos_env__$private$pos[indx]
@@ -149,7 +157,7 @@ makeUR <- function(RAobj, filter=list(MAF=0.05, MISS=0.5), ploid=2, mafEst=TRUE,
 
   ## Update the R6 objective
   URobj$.__enclos_env__$private$updatePrivate(list(
-    genon = genon, ref = ref, alt = alt, chrom = chrom, pos = pos,
+    genon = genon, ref = ref, alt = alt, chrom = chrom, pos = pos, nInd = length(indID),
     SNP_Names = SNP_Names, nSnps = nSnps, AFrq = AFrq, pfreq = pfreq, ep = ep)
   )